Create cssfront.min.css
This is the combination of cf-foundation.css and cf-decoration.css. Minified for use in production sites. Please include css-front files via rawgit.com's CDN. This file can be included with this link in the head part of an HTML file: https://cdn.rawgit.com/ox-harris/cssfront/master/src/cssfront.min.css. Please visit rawgit.com for details on how these links are generated. And thanks rawgit.com for the service.